+++ quantforge/__init__.py
@@ -521,6 +521,7 @@ from .differential_evolution import differential_evolution
 from .nnls import nnls
 from .simulated_annealing import simulated_annealing
 from .bfgs import bfgs
+from .conjugate_gradient import conjugate_gradient, gauss_seidel, jacobi
 from .levenberg import levenberg_marquardt
 from .ewma_cov import ewma_covariance_matrix, ewma_correlation_matrix
 from .vol_convert import black_to_normal_vol, normal_to_black_vol
@@ -911,7 +912,7 @@ try:  # pragma: no cover - trivial availability branch
 except ImportError:  # pragma: no cover
     HAS_NUMPY = False
 
-__version__ = "1.935.0"
+__version__ = "1.936.0"
 
 __all__ = [
     "OptionType",
@@ -1674,6 +1675,9 @@ __all__ = [
     "nnls",
     "simulated_annealing",
     "bfgs",
+    "conjugate_gradient",
+    "gauss_seidel",
+    "jacobi",
     "levenberg_marquardt",
     "ewma_covariance_matrix",
     "ewma_correlation_matrix",